feat: uuidの差し替えinterfaceをcoreパッケージに切り出し
